* Refactored xml stuff.
