+++ src/main/java/org/openpnp/machine/reference/psh/NozzlesPropertySheetHolder.java
@@ -51,6 +51,7 @@ public class NozzlesPropertySheetHolder extends SimplePropertySheetHolder {
             }
             try {
                 Nozzle nozzle = cls.newInstance();
+
                 head.addNozzle(nozzle);
             }
             catch (Exception e) {
